[ui] Re-focus the original path after editing is completed
When editing things like a combo in the memory editor, accepting the
edit will screw up the tab order by choosing something else to focus
after the ephemeral widget used to do the edit is removed. This saves
the current path when editing starts and restores it when editing
is finished to avoid messing up the workflow.

Fixes #4865