Some fixes to run_tests
First, in r2787 we made the BruteForce test do RadioFeatures validation on
a memory, avoiding a failure if the memory did not validate. For some reason
we still required that radio.validate_memory() pass for any of those tests,
which doesn't allow for the radio to ban quirks in the tests. This makes us
not fail if the radio rightly fails to validate.

Second, we were modifying the memory in do_mode() so that once we passed FM
in the mode sequence we were always trying freq+100Mhz. Depending on how we
ran the tests (i.e. which order we hit the modes) this would sometimes fail.
This makes us copy the original memory so we're stable throughout the modes.

Related to #5237