[PATCH][FD-268] Match some LCD icons features and Chirp way in settings, fixes #5733
Some of the LCD icons has an associated feature in the settings part,
now the driver match that behaviour.

Also the settings part in the driver is more in chirp's way of doing it, this
was my frist driver and I made it different in that time.

PEP8 compatibility at 99% also included on the driver update now.

73, Pavel, CO7WT.